% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_definitions.R
\name{custom_themes}
\alias{custom_themes}
\alias{theme_nyt}
\alias{theme_midnight}
\alias{theme_royal}
\alias{theme_deepblue}
\title{Custom Plot Themes}
\usage{
theme_nyt(gridline_x = TRUE, gridline_y = TRUE)

theme_midnight(base_size = 11, base_family = "")

theme_royal(base_size = 11, base_family = "")

theme_deepblue(base_size = 11, base_family = "")
}
\arguments{
\item{gridline_x, gridline_y}{For \code{theme_nyt()}: if \code{TRUE} (the default), draw
the major gridlines of the x or y axis as thin dashed grey lines;
otherwise omit them.}

\item{base_size, base_family}{For \code{theme_midnight()}, \code{theme_royal()} and
\code{theme_deepblue()}: base font size (in points) and font family, passed on
to \code{\link[ggplot2:theme_gray]{ggplot2::theme_gray()}}.}
}
\value{
A \code{ggplot2} theme object.
}
\description{
Custom themes included with the package that can be added directly to a
\code{ggplot2} plot.
}
