% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_definitions.R
\name{ggplot2_md_themes}
\alias{ggplot2_md_themes}
\alias{md_theme_gray}
\alias{md_theme_grey}
\alias{md_theme_bw}
\alias{md_theme_linedraw}
\alias{md_theme_light}
\alias{md_theme_dark}
\alias{md_theme_minimal}
\alias{md_theme_classic}
\title{Markdown-Enabled ggplot2 Themes}
\usage{
md_theme_gray(...)

md_theme_grey(...)

md_theme_bw(...)

md_theme_linedraw(...)

md_theme_light(...)

md_theme_dark(...)

md_theme_minimal(...)

md_theme_classic(...)
}
\arguments{
\item{...}{Additional arguments passed to the original theme function.}
}
\value{
A markdown-enabled \code{{ggplot2}} theme object.
}
\description{
A collection of standard \code{{ggplot2}} themes with markdown support enabled.
These can be added to a plot using the \code{+} operator.
}
\details{
The titles, captions, axis titles, legend titles and facet strips render
markdown. The tick labels of the axes and the labels of the legends stay plain:
they come from your data, and a label such as \verb{<LOD>} would be read as markup.
Markdown legend labels would also make a continuous color bar as long as the
canvas is high. To use markdown in the tick labels or in the labels of a
discrete legend, switch the element on for that plot, for example
\code{theme(legend.text = ggtext::element_markdown())}.
}
