% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_set.R
\name{list_color_scales}
\alias{list_color_scales}
\title{List Available Color Scale Sets}
\usage{
list_color_scales()
}
\value{
A character vector of available scale set names.
}
\description{
Prints the names of the exported discrete color and fill scale sets.
}
