% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_set.R
\name{list_theme_sets}
\alias{list_theme_sets}
\title{List Available Theme Sets}
\usage{
list_theme_sets()
}
\value{
A character vector of available theme set names.
}
\description{
Prints the names of the theme sets available to use with \code{apply_theme_set}.
}
