% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_set.R
\name{reset_global_theme}
\alias{reset_global_theme}
\title{Reset the Global ggplot2 Theme and Scales}
\usage{
reset_global_theme()
}
\description{
Resets the global theme to \code{theme_gray()} and unsets the global default
color and fill scales, restoring the ggplot2 defaults.
}
