% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_definitions.R
\name{theme_color_scales}
\alias{theme_color_scales}
\alias{scale_color_npg}
\alias{scale_fill_npg}
\alias{scale_color_jama}
\alias{scale_fill_jama}
\alias{scale_color_avatar}
\alias{scale_fill_avatar}
\alias{scale_color_simpsons}
\alias{scale_fill_simpsons}
\alias{scale_color_olink}
\alias{scale_fill_olink}
\alias{scale_color_fivethirtyeight}
\alias{scale_fill_fivethirtyeight}
\title{Discrete Color and Fill Scales}
\usage{
scale_color_npg(...)

scale_fill_npg(...)

scale_color_jama(...)

scale_fill_jama(...)

scale_color_avatar(...)

scale_fill_avatar(...)

scale_color_simpsons(...)

scale_fill_simpsons(...)

scale_color_olink(...)

scale_fill_olink(...)

scale_color_fivethirtyeight(...)

scale_fill_fivethirtyeight(...)
}
\arguments{
\item{...}{Additional arguments passed to the underlying scale function.}
}
\value{
A \code{ggplot2} scale object.
}
\description{
Convenience wrappers around palette scales from supporting packages. These
functions can be added directly to a \code{ggplot2} object.
}
\seealso{
\code{\link[=scale_color_set]{scale_color_set()}} and \code{\link[=scale_fill_set]{scale_fill_set()}} to pick any palette by name.
}
